pyproject.toml
[tool.poetry]
name = "psyncly"
version = "0.1.0"
description = ""
authors = ["felixcapuano <felix.capuano@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"


[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = 'scripts:dev'
fmt = 'scripts:format'
